CachingBuildEngine: Extract a method object for buildLocally
Summary:
This commit extracts `buildLocally` into a new class, `BuildRuleSteps`. Having the complete
task (including tracing and everything) encapsulated in an object will make it more straightforward
to ensure that the same code continues to run for a given rule as we start to build multiple rules
in a single submission to the `ExecutorService`.

Test Plan: CI

fbshipit-source-id: b417607