CachingBuildEngine: Make `BuildRuleSteps` provide its own `Future`
Summary:
Instead of `submit`ing `buildLocally` to the `ExecutorService` as a `Callable`, we pass it the
`ExecutorService`, which it in turn uses to `execute` the `BuildRuleSteps`, returning the `Future`
provided by `BuildRuleSteps`.

By allowing `buildLocally` to determine what `Future` will be used to detect completion of a particular
rule, we make it possible to build multiple rules from a single `ExecutorService.execute` call.

Test Plan: CI

fbshipit-source-id: ab74950