Add instructions on sharing macros to docs
Summary:
I was recently trying to get some macros setup for a project and I couldn't remember how includes were supposed to work, and it would have been helpful if the page on macros was more instructive. This is just adding a section to introduce the concept of project includes so it's easier to come across/is in more than one place.

Also add some section headers to break up the article.
Closes https://github.com/facebook/buck/pull/1065

Reviewed By: ttsugriy

Pulled By: ttsugriy

fbshipit-source-id: f54d673