elf: add step which rewrites the dynamic string section
Summary:
This adds a new step which will rewrite the dynamic string table with the
strings references by given dynamic sections.  The idea is to run this after
scrubbing certain components of dynamic sections to compact the dynamic string
section and drop unused strings.

Test Plan: added a simple test

Reviewed By: yiding

fbshipit-source-id: ef59f37