elf: split `ElfExtractSectionsStep` into more granular steps
Summary:
This splits the current `ElfExtractSectionsStep` into two separate steps for
extracting particular steps and for compacting the current sections of an ELF
file.  This is a noop for now, but will become necessary for future changes that
will need to perform additional operations between these two steps.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 7c59f3c