CxxPreprocessAndCompile: desandboxing input when calculating getInputsAfterBuildingLocally
Summary:
The current logic of handling depfiles for cxx requires that all headers files and source
files are "deciphered' from links from symlink tree to original locations.
It already happens in Depfiles.parseAndOutputBuckCompatibleDepfile via header normalizer.
This special case is for applying the same logic for an input cxx file in the case
when cxx.sandbox_sources=true.

Test Plan: CxxDependencyFileIntegrationTest is ok now with `buckd` and `sandboxSource`.

fbshipit-source-id: 8846910