python: don't require that python platform intepreters exist
Summary:
Previously, when constructing `PythonPlatform`s, we'd fail if the path referred
to by the sections `interpreter` field didn't exist as an executable file.  This
is unnecessary as we don't actually need this path available to build a python
binary using it (e.g. it can just be deployed to the host that runs it).  Also,
this made it difficult to setup a `.buckconfig` to define several python platforms,
where only a subset were valid for a given host type.

There were two operations that did actually require the interpreter be available:
1) To extract the python version.  To work around this, this diff adds a new
   `version` field so a python platform section can specify an explicit version
   and avoid having to attempt to execute the interpreter.
2) For parsing (as the parse calls into `PythonBuckConfig` methods to grab the
   path to an operable interpreter).  This diff spins that functionality off
   into a separate `getRunnablePythonInterpreter` method.

Test Plan: CI

Reviewed By: bertmaher

fbshipit-source-id: afb09b3