Don't delete linker argsfiles after linker step runs
Summary:
Having these get emitted into the log at -v 9 was not a good solution as it
requires people to manually find and copy paste sections of those logs into
files in order to be able to manually rerun commands.

This behavior is also inconsistent with the lack of clenaup for other
argsfiles, such as for compilation, and those are much bigger anyway.

Test Plan: ran a simple link job and see that the scratch files are present after the build.

Reviewed By: dinhviethoa

fbshipit-source-id: a254c42