fix RandomizedTrial to pick properly
Summary:
Simplify the logic of picking a random double. Instead of doing some arbitrary
walk inside a hash, just take the first 8 bytes and seed a random generator, of
which we ask for a double.

This fixes the problem that the value was previously quantized to 0/255.0
through 255/255.0.

Additionally, the returned value is 0 to 1 exclusive instead of inclusive. This
fixes the problem that the last element of a set is never picked when the
result is 1, which used to occur 1/255 of the time.

Since the picking logic is now correct, we don't need this stupid hack to
return a default value if picking fails, so I removed it.

Also simplify the picking logic code to just keep subtracting the picked value
until it's less than zero.

Test Plan: The test are still kind of terrible, but they probably wouldn't fail now.

fbshipit-source-id: c2fb791