Throw IllegalStateException for package boundary exception
Summary:
I don't see any way to hit this error that doesn't involve a bug in buck.

- We can't refer to a file in a parent or sibling directory because of the above
  check (unless we have package boundaries turned off, in which case we can't
  hit this exception anyways).
- If the directory containing the build file is ignored, you'll get a message about
  how that target can't be built because it is in an ignored directory
- If a file in an ignored subdirectory is referenced, the build still succeeds.

Turning it into an IllegalStateException lets us catch it in our logging tools,
which will make it easier to find and fix bugs in buck.

Test Plan:
introduce a trivial bug so this actually triggers, and observe the
resulting stack trace

fbshipit-source-id: 1d21829