windows cxx: lib.exe @ar.argsfile
Summary:
`Ar` utility on *nix first creates an archive if it doesn't already exists, and then adds files to it.
On the contrary, MS `Lib.exe`  on windows, always creates a new archive.
So, to be able to create a large lib, an argsfile is used now on windows.
(More precisely: it is possible to use the same logic on windows, but it is confusing:
to create an archive: `lib.exe /out:archive.lib m1.obj m2.obj`,
to add to the archive: `lib.exe archive.lib m3.obj m4.obj` - this difference
in syntax is not documented by MS).

Test Plan: CI: WindowsCxxIntegration test is green

Reviewed By: mzlee

fbshipit-source-id: e805330