Allow named root cell config overrides
Summary:
Sometimes we want to override a config with the root cell
name. Add parsing in CellConfig to handle this case and have the Main
initialization resolve this case.  This means the root cell is allowed
two names: empty and non-empty and the configs are merged.

Test Plan: Add a new case in the unit test

Reviewed By: yiding

fbshipit-source-id: d30116f