Jsr199Javac: Fix another crashy compiler error case
Summary:
Invoking the several stages of `JavacTask` one after another does not appear to be a well-tested
scenario. This commit addresses another issue that's been found with doing this -- even with our previous
fix of checking for errors between the phases, the compiler can swallow legitimate errors and crash
instead later on. From code inspection, it looks like `javac` 9 would just spew extra errors in this case,
but `javac` 8 crashes. The fix is to force errors to be reported between the phases.

Test Plan: New integration test

Reviewed By: dreiss

fbshipit-source-id: 0e680fd