Clarify package boundary message
Summary:
The message current assumes that the file is already referenced by
another rule in the owning BUCK file, but that's not true (and could be
really confusing when it's not true). This adds some text about the
'owning' BUCK file without reintroducing the "package boundary" concept.

It also gets the build file name correctly from the target cell object.

Adds back the suggestion to refer to the owning rule instead of the file
directly.

Old:
```
The target '//java:foo' tried to reference 'java/dir/Foo.java'.
It's not allowed because this file is already referenced by another rule
in 'java/dir/BUCK'.

It may also be due to a bug in buckd's caching.
Please check whether using `buck kill` will resolve it.
```
New:
```
The target '//java2:foo' tried to reference 'java2/dir/Foo.java'.
This is not allowed because 'java2/dir/Foo.java' can only be referenced from 'java2/dir/BUCK'
(which is it's closest 'BUCK' file).

You should find (or create) the rule in 'java2/dir/BUCK' that references
'java2/dir/Foo.java' and use that in '//java2:foo'
instead of directly referencing 'java2/dir/Foo.java'.

This may also be due to a bug in buckd's caching.
Please check whether using `buck kill` will resolve it.
```

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: fc1c825