Remove enhancementResult field from AndroidBinaryBuildable
Summary:
This extracts all the parts of the enhancementresult that are accessed
into their own fields (and derives a couple values). Some of this can
already be added to the rulekey, the rest continue to not be added, but
it will be easier to incrementally fix them.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 97541fb