Add support for spelling corrections to subcommands
Summary: Adds support to our buck subcommand parser to be lenient on small spelling errors.

Test Plan:
```
$ ./bin/buck trgets buck
Assuming 'trgets' is a creative spelling of 'targets'
//programs:buck

$ ./bin/buck targetss buck
Assuming 'targetss' is a creative spelling of 'targets'
//programs:buck

$ ./bin/buck targetd buck
Assuming 'targetd' is a creative spelling of 'targets'
//programs:buck

$ ./bin/buck targe buck
"targe" is not a valid value for "subcommand"
For help see 'buck --help'.
```

Reviewed By: dinhviethoa

fbshipit-source-id: 8160278