Clear metadata from DB before filesystem
Summary:
The DB contains things like RULE_KEY that are used for determining
cache hits, which then require RECORDED_PATHS to be present.  If something has
caused the process to die between removing RECORDED_PATHS and removing the other
metadata, we end up stuck in a manner that requires `buck clean`, because
RECORDED_PATHS is gone.

If we always remove metadata from the DB first, this can't happen.

Test Plan: thought experiment

Reviewed By: dinhviethoa

fbshipit-source-id: 8dd53a0