haskell: don't cache thin archives
Summary:
As per the comment in this diff, currently, GHC produces nono-deterministically sized
object files.  This means that it's possible to get a thin archive fetched from cache
originating from one build and the associated object files fetched from cache originating
from another, in which the sizes listed in the archive differ from the objects on disk,
causing the GHC linker to fail.  Technically, since the `HaskellCompileRule` is producing
non-deterministic outputs, we should mark that as uncacheable.  However, as that would
have a significant affect on build efficiency, and since this issue appears to only
manifest by a size mismatch with what is embedded in thin archives, just disable caching
when using thin archives.

Test Plan: CI

Reviewed By: JonCoens

fbshipit-source-id: 1b60c3b