Migrate CompileToJarStepFactory family from RuleKeyAppendable to AddsToRuleKey
Summary:
This commit does a direct migration of what was in the `appendToRuleKey` methods
to `AddToRuleKey` annotations. In a couple of cases `appendToRuleKey` chained to
another class's `appendToRuleKey`...in those cases there's now an extra level of
indirection in the rule key, but it will still change under the right circumstances.

Looking at the class definitions after this commit, it seems likely that some of
the fields that aren't part of the rule key technically should be. This commit is just
a straight migration and doesn't attempt to deal with that. We haven't seen problems with
these yet, probably because the fields are marked `AddToRuleKey` in the owning rule itself
or another helper object.

Test Plan: Code inspection, CI

Reviewed By: dreiss

fbshipit-source-id: 8fdcc6e