Revert "CachingBuildEngine: Separate getting Steps from running them"
Summary:
Generating steps at enqueue-time instead of execution-time causes the Step
objects to be held in memory for a longer period of time, increasing memory
usage.

CxxPreprocessAndCompileStep, in particular, holds a HeaderPathNormalizer
instance. These instances can be tends of megabytes each, and there's one per
cxx_library. If compile rules from multiple libraries are enqueued
simultaneously, these instances can use up gigabytes of heap. This was not a
problem prior to this change as HeaderPathNormalizers are weakly held in
`PreprocessorDelegate` and re-created when necessary.

Test Plan: Build many static libraries in parallel.

Reviewed By: dinhviethoa

fbshipit-source-id: e6c0698