Compute some things in constructor of AndroidBinaryBuildable
Summary:
This computes a boolean, 'isPredexed' in the constructor and then makes
addFinalDxSteps have an early exit for that case. Changes the inputs
unique to the rest of addFinalDxSteps to be Optionals (that are
explicitly set to empty() for predexed cases). This allows us to make
those inputs AddToRuleKey (previously, those lists would include a
bunch of intermediate dex inputs that aren't actually needed as inputs
in the predexed case).

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: e468269