Fix bisection edgecase: paths that are invalid prefixes
Summary:
When bisection finds a path that is an invalid prefix, like

    foo/bar-spam/

for the target

    foo/bar

then String.compare() will produce the //wrong// result (in that Java would
want the shorter prefix to be sorted before the longer path). Fix by doing the
comparison with `dirname` (the path with `/` appended).

Test Plan:
Added a new test that reproduces the error; this diff successfully passes that
test.

Reviewed By: yiding

fbshipit-source-id: 53fa21b