CachingBuildEngine: Finalize rules immediately after their build, on the same thread
Summary:
Previously when the main part of building a rule completed, we'd enqueue another task to "finalize"
the rule (run post-build steps and the like). Only after the finalization completed would we release
the dependents to build. Performing post-build steps right afterwards on the same thread makes it simpler
to pipeline rule execution without causing the order of execution to change.

Test Plan: CI

fbshipit-source-id: 1b8af12