Remove the supplier of secondary dex sourcepaths
Summary:
This moves the logic in PreDexMerge around so that this list of
directories isn't written to disk and then loaded in initializefromdisk
(it's trivial to determine the list ahead of time).

Then, AndroidBinaryBuildable can hold this in a field that is actually
added to the rulekey.

At this point, all the inputs to AndroidBinaryBuildable are added to the
rulekey and we can start dismantling the ComputeExopackageDepsAbi stuff!

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: aada3df