Make AndroidBinary acquire the primary dex explicitly
Summary:
The way that the predexed primary dex was getting to the
AndroidBinaryBuildable was odd (basically, PreDexMerge asked the
AndroidBinary where to put it). The information should flow bottom-up,
not top-down.

This changes PreDexMerge to just put the primary dex whereever it wants,
and AndroidBinaryBuildable gets a SourcePath to it (or creates its own
primary dex in its own location if not predexed).

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 10c2b05