Add ModernBuildRule scaffolding
Summary:
This adds the Buildable interface and ModernBuildRule. It adds a bunch
of the interfaces around this and some of the simpler implementations.

ModernBuildRule is an implementation of BuildRule that greatly limits
the implementation of the parts of BuildRule that the CachingBuildEngine
cares about.

A ModernBuildRule takes a Buildable as one of it's constructor params.
Most of the BuildRule implementation is actually based on that
Buildable. getBuildSteps() is implemented as a small wrapper around a
more limited version of getBuildSteps() on the Buildable.

A Buildable is limited (or rather, will be limited once the
implementation is complete) to being a top-level class or static inner
class with only final fields of a limited set of types.

A couple of new types (Input/OutputPath and Input/OutputData) are
introduced for Buildable fields. These allow ModernBuildRule and the
CachingBuildEngine (and maybe others) to better understand and restrict
the flow of data.

Deps, outputs, and rulekey are all derived simply from the Buildable's
fields.

Test Plan: .

Reviewed By: bertmaher

fbshipit-source-id: d782499