Small refactoring in AbstractInvocationInfo to improve information hiding.
Summary:
It felt very weird to me that `AbstractInvocationInfo` had a package-private
static member named `DIR_NAME_REGEX` that it did not use. Looking closer, it
had an implicit relationship with `DIR_NAME_TEMPLATE`, which is presumably why
they were declared together. (Though this relationship was undocumented, which
I fixed in this commit.)

Further, `DIR_NAME_REGEX` was package-private so that `LogFileHandler` could use
it. In practice, `LogFileHandler` does not really need to know about the value
of `DIR_NAME_REGEX`: it only cares whether a path is one that was minted by
`AbstractInvocationInfo`. As such, I think that knowledge should be encapsulated
in `AbstractInvocationInfo`, so I added `AbstractInvocationInfo.isLogDirectory()`
to achieve that. Now `DIR_NAME_REGEX` has been replaced by `DIR_PATTERN`, which
is private rather than package-private.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 64d06db