CachingBuildEngine: Support pipelining steps from multiple rules (take 2)
Summary:
This is a second (less hacky) attempt at supporting pipelining of rules.

Some rules can benefit from greater efficiency if their steps are run immediately
after those of a dependency. For example, a `java_library` rule can re-use the `javac`
instance that was used to construct its ABI.

This commit introduces the support for such situations in `CachingBuildEngine`. Rules
that need pipelining implement `SupportsPipelining`. When such a rule is built locally,
we build all rules in its pipeline back-to-back on the same thread. Each rule is built with
its own `BuildableContext` and appropriate tracing, so the rest of the system should not
see any difference beyond the change in scheduling.

Test Plan: New unit tests

fbshipit-source-id: 183a826