DefaultJavaLibraryBuilder: Create ABI rules up front
Summary:
This commit changes `DefaultJavaLibraryBuilder` to create the ABI rules at the same time
as it creates the library rules. This allows the ABI and library to share a `JarBuildStepsFactory`
instance. Sharing the instance makes pipelining the rules together more obviously correct, since
all of the settings that go into creating the compiler come from this factory. It has the side
benefit of making action graph creation slightly more efficient (or at least making it not require caches
for efficiency in this particular case).

The changes in `JarBuildStepsFactory` are a quick hack to work around the statefulness of `JavacToJarStepFactory`;
subsequent commits will fix that up.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 852a8ec