Use ImmutableList<String> instead of String[] for args.
Summary:
Although the standard Java `main()` receives args as a `String[]`, that doesn't
mean we should perpetuate this throughout our code. `String[]` is more difficult
to deal with for a number of reasons.

The one that I am currently running into is that when it's used as a value in a
`Map`, I cannot `assertEquals()` using such maps in a unit test because
`.equals()` is not defined as we want for arrays, but it is for `List`.

Normally, you use `Assert.assertArrayEquals()` from JUnit to work around this,
but it's not so easy to do when the `Map` has heterogeneous value types because
it is representing JSON. I'm not keen on writing a recursive assertion method
for this case when switching to `List` seems cleaner, anyway.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: c3d338c