Update get_jars.sh with stubbed android.jar for FeatureEnableService
Summary:
This updates the `get_jars.sh` script to pull and stub the android.jar from an IntelliJ instance. The ` /Applications/IntelliJ IDEA.app/Contents/plugins/android/lib/android.jar` is needed as a provided dependency for #1434 but will ultimately be supplied by the IntelliJ Android plugin dependency. It does not need to be a specific version.

    Version/build info from /Applications/IntelliJ\ IDEA\ CE.app/Contents/lib/resources.jar!idea/IdeaApplicationInfo.xml
          <version codename="Community Edition" major="2017" minor="1.5" eap="false"/>
          <build number="IC-171.4694.70" date="201707040805"/>

Closes https://github.com/facebook/buck/pull/1448

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: d608c85