Log network info asynchronously
Summary:
On Mac OS, We do a lot of calls to `networksetup` to get the active
network medium and SSID, and these can be slow (especially if you have a lot of
devices and your WiFi is low in the service order... which is how I found this).

To avoid putting this on the critical path, let's do it on a service thread and
post the result back when it's ready.  We'll lose network data for fast
commands, but that seems better than making users wait for it.

Test Plan:
`buck root` goes from ~3 seconds to ~1 on my MacBook.  Logging is
unaffected for sufficiently long-running commands.

Reviewed By: dinhviethoa

fbshipit-source-id: c74a489