Include codesign provider target as apple_test dependency
Summary:
It's possible to have a target graph that includes an apple_test but no apple_bundle. However, apple_test ends up creating an apple_bundle as part of it, without getting runtime deps from the AppleBundleDescription. If codesign is configed to be a target (and not just a binary) with this setup, the codesign target is never added to the target graph so the provider fails to look it up when codesigning the bundle.

This seems like a deeper problem of creating the AppleBundle through the AppleDescriptions factory method rather than properly depending on AppleBundleDescription in AppleTestDescrption, but the solution in this diff is analogous to existing `apple.xctool_zip_target` and unblocks me from using a custom target for codesigning.

Reviewed By: ryu2

fbshipit-source-id: b3bf5d5