Make CachingBuildRuleBuilder always use delegate.shouldKeepGoing
Summary:
There was one instance where this builder was using custom logic to
determine if it should keep going. Change this to use shouldKeepGoing()
everywhere.

The only difference is that in the old code, this would break early if
using --populate-cache and there is an error, now it will continue.

Test Plan: CI

Reviewed By: bertmaher

fbshipit-source-id: 6ef77e7