Add a class to manage BuildRule scopes.
Summary:
Right now, BuildRuleEvent.resumeSuspendScopes cannot be nested on the
same thread. This means that, to correctly keep everything in a
buildrule scope we must:
1. Usually, when doing Futures.transform/etc, we must wrap the transform
in a scope.
2. But, if it's actually running on a direct executor, it must not be
wrapped in a scope.

This adds a simple class that allows nested buildrule scopes and
disallows having multiple threads be within the same buildrule's scopes
(that's not something we do right now, and we have no need for it).

Test Plan: CI

Reviewed By: jkeljo

fbshipit-source-id: decc456