Suppress PROCESSING FILESYSTEM CHANGES when it's fast
Summary:
This message shows up on basically every buck command and it's often
some trivial amount of time like 0.2 seconds.  Since the point is to reassure
users in the (rare) case when this takes many seconds, let's not spam the
console unless it's been at least 1 second.

I added this to the logEventPair API since (a) it wasn't too hard, and (b) I
could see us wanting to use this for other events to provide a non-spammy
progress meter.

Test Plan: Observe superconsole not printing this message when it's fast.

Reviewed By: ttsugriy

fbshipit-source-id: ec87c50