Add LimitedFileHashCacheEngine
Summary:
This adds a filehashcache engine that only caches values on non-symlink
files. This means that it handles symlinks and directories containing
symlinks properly.

With this hashcache engine, we could have rules actually use the outputs
of symlinktrees without needing to know that they are referring to
symlink trees (right now, to be input-based, they need to add the
symlink tree paths to their rulekeys as nonhashablesourcepathcontainer
and add the symlink target as a sourcepath).

The underlying implementation is the new prefix-tree FileSystemMap.

Unlike other engines, this one only has a single underlying cache
instead of two (other engines have one for size, one for
hash+othermetadata).

Test Plan: CI, the main filehashcache tests run over all filehashcachemodes

Reviewed By: bertmaher

fbshipit-source-id: b1ff16c