Add BuildableSupport.deriveDepsFromAddToRuleKeys
Summary:
This function essentially derives a list of deps from everything in the
rule that is added to the rulekey. For input-based rules that specify
their rulekeys correctly, this should be exactly the right set of deps.
A lot of rules were deriving there deps in this way already (but
hard-coding the things to derive deps from).

Test Plan: added test

Reviewed By: yiding

fbshipit-source-id: 4e3ae74