Provide a way to freeze global variables.
Summary:
One of the reasons for hash and rulekey non-determinism is
mutations of global variables. This can also be mitigated by using copy.deepcopy
but it can lead to developer confusion. I realize that even freezing variables done
in the way it's currently implemented can lead to frustration, since a failure
would be saying that the type developers expect like list does not have append method.
To mitigate this we can write our own immutable list implementation, where append
method would throw an exception clearly stating that `append` method is not allowed
on a frozen variable.

Test Plan: ant buck-py-test

Reviewed By: yiding

fbshipit-source-id: 846a0f4