Wire up freeze_globals functionality
Summary:
Previously `freeze_globals` flag was no configurable and impossible to
use when running buck.
If attempt to modify a global list is made and globals are frozen parser
will throw an exception with text like
```
'tuple' object has no attribute 'append'
...
FOO.append('bar')
```
This is a far from being perfect message, since it's not obvious why
`FOO` suddenly became a `tuple`, and as a follow-up I'm considering:
- adding an error converter to explicitly point out what's going on
- moving from python to Skylark, where semantics of the language can
be controlled much better. For now I would like to use this as a way
to prevent common way of non-determinism from being introduced into
the codebase.

Test Plan: buck test ...

Reviewed By: dinhviethoa

fbshipit-source-id: e089e60