Add ability to pass -swift-version for Swift
Summary:
In order to able to specify the Swift version on a per-target basis, we need to be able to pass the `-swift-version` parameter to `swiftc`. This diff makes the following changes:

- Adds `swift_version` to `apple_library`
- Adds `version` to `swift_library`

In the case of `apple_library`, `swift_version` is used in two places:

- Version to pass to `swiftc`. Do note that `swiftc` accepts only major version, so the version specified in the BUCK file gets truncated appropriately.
- During Xcode project generation to specify `SWIFT_VERSION` build setting. If `apple_library` does not specify an explicit `swift_version`, project generation falls back to previous behaviour of picking up the `version` field from the `swift` section in the Buck config.

This change enables the use of distinct Swift language versions for different modules which is a part of [Swift 4](https://swift.org/blog/swift-4-0-release-process/).

This change is designed to be backwards compatible and one of the reasons to not use the Swift `version` field from the Buck config in the construction of the arguments to `swiftc`.

Do note that the Buck config's Swift `version` field is still used to look up the toolchain.

Test Plan: Unit tests

Reviewed By: dinhviethoa

fbshipit-source-id: 99bb517