Switch from defaultCxxPlatform to defaultCxxFlavor
Summary:
defaultPlatform is actually a defaultFlavor in disguise in a lot of places, so change that in preparation for more refactoring.

The eventual goal is to be able to have all cxx-related library, test, and binary targets specify their own default flavor (e.g., android-armv7) so that, when compiling just that target, it gets the correct CxxPlatform. Right now, an apple_binary does not allow specifying a default_platform directly, so we need to go the round-a-bout method of `--config cxx.default_platform=iphonesimulator-x86_64`.

Test Plan: Update unit tests

Reviewed By: andrewjcg

fbshipit-source-id: ea05092