Make manifest-based cache fetch async
Summary:
The goal here is to move the posting to the cacheActivityService
down into the cache itself. First, we'll move it down into
CachingBuildRuleBuilder.fetch(), and then into the cache.

This makes the manifest fetch and the following artifact fetch async.
This is still on the normal service (but the next diff will move it).
This is the last use of the underlying fetch() that forced it to be
synchronous.

Test Plan:
CI
`buck build -c mode=dir,thrift_over_http buck` verify superconsole and chrome trace look correct

Reviewed By: dinhviethoa

fbshipit-source-id: d69418b