Create helper method for submitting a metadata event to a Chrome trace.
Summary:
This fixes an issue where the timestamp was erroneously set to zero.
Although `microThreadUserTime` is not quite set correctly, this seems better
than what we had. Incidentally, the `tts` field is optional:
https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview

This new method will also be used in a subsequent change to include Watchman
events of interest to the Chrome trace.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: b94a675