Fix formatting of ideabuck change notes.
Summary:
As observed on multiple occasions, the change notes for the
ideabuck plugin are half-markdown, half-html, and so they
read confusingly.  Fixed by making it all-html.

Test Plan:
Build and install locally from the buck repo.
```
$ bin/buck build ideabuck --show-output
...
The outputs are:
//src/com/facebook/buck/intellij/ideabuck:ideabuck buck-out/gen/src/com/facebook/buck/intellij/ideabuck/ideabuck.jar
$ cp buck-out/gen/src/com/facebook/buck/intellij/ideabuck/ideabuck.jar ~/Library/Application\ Support/IdeaIC2017.1/
```
Run IntelliJ CE 2017.1 and using `cmd-,` select plugins, select `Buck for IDEA` and observe the Change Notes.

Reviewed By: styurin

fbshipit-source-id: 768050e