Simplify AppleDebuggableBinary
Summary:
It used to be that to create one of these rules, one would call a static
method, then pass the result back into the constructor. Instead of that, create
factory functions for the three scenarios.

Get rid of input based rulekey interface, since this builds nothing so is
unlikely to benefit.

Inherit directly from AbstractBuildRule, as BuildRuleParams doesn't really add
anything here.

Allow dsym to be retrieved, for action graph construction purposes (AppleBundle
creation needs this). This used to be re-computed, now it can simply be
retrieved from this rule instead.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: fb79a86