haskell: use `Supplier`s for GHCi config settings
Summary:
This switches to using a `Supplier<Path>` for the various GHCi config settings
so that calls to the `HaskellBuckConfig` methods can be made eagerly without
failing hard when these aren't set (for builds that don't use them).

This is a prereq to moving to multi-haskell builds, as `HaskellBuckConfig` will
move to a eagerly built `HaskellPlatform`.

Test Plan: CI

Reviewed By: JonCoens

fbshipit-source-id: 8cda218