RemoveClassesPatternsMatcher: Implement Predicate for ease of use
Summary:
Implementing `Predicate` enables `JarParameters` to more easily work with both libraries (which use
`RemoveClassesPatternsMatcher`) and binaries (which use a `Predicate`).

This does end up requiring some casting in places that know they're dealing with a library (such
as the out-of-process `javac` marshalling code), but overall I think it's a good trade.

Test Plan: CI

fbshipit-source-id: b19b2ad