haskell: use `HaskellPlatform` flavor instead of `CxxPlatform`
Summary:
Now that we're threading `HaskellPlatform` throughout haskell rule generation,
instead of `CxxPlatform`, use the former's flavor to name build rules.  We still
currently require that `HaskellPlatform.getFlavor() == `CxxPlatform.getFlavor()`,
as we rely on this to lookup the `HaskellPlatform` from the `CxxPlatform`
passed to the various C/C++ interfaces that haskell rules implement.

Test Plan: CI

Reviewed By: JonCoens

fbshipit-source-id: a57ac8f