Better handling of empty stderr section
Summary:
The test for parsing out stderr information from a Mercurial traceback produced
a newline too many when that section is empty:

```
stderr:

Traceback ...
```

In reality that newline is not there, and there was an off-by-one error in the
code parsing this out. Fix the test and the error.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 45fac5f