Introduce [eden] .buckconfig option: disable_sha1_fast_path.
Summary:
`EdenProjectFilesystemDelegate` has a special implementation of
`computeSha1(Path)` to request the SHA-1 from Eden via Thrift rather
than reading the contents of the file and computing the SHA-1 itself.

Although we are optimistic that this will be an overall performance win,
we need to back that up with data, so making this behavior configurable
will make it easier to profile.

Test Plan:
Manually verify that toggling the option has the intended behavior.

Made a local change to add some logging to `EdenProjectFilesystemDelegate`:

```
    if (disableSha1FastPath) {
      LOG.verbose("Computing SHA-1 for %s", path);
      return delegate.computeSha1(path);
    } else {
      LOG.verbose("Fetching SHA-1 for %s", path);
    }
```

Ran `buck build` in an Eden mount and verified I saw a bunch of messages like:

```
[2017-08-13 22:49:51.507][vrbos][command:d67c55dd-f471-43b3-aacf-703c8bfd5d17][tid:250][com.facebook.buck.eden.EdenProjectFilesystemDelegate] Fetching SHA-1 for <absolute path>
```

Then I added the following to my `.buckconfig.local`:

```
[eden]
  disable_sha1_fast_path = true
```

and now when I ran the command, I saw:

```
[2017-08-13 22:47:19.561][vrbos][command:9e4e71d5-4eb4-4eb4-9443-a5fe4bf611eb][tid:119][com.facebook.buck.eden.EdenProjectFilesystemDelegate] Computing SHA-1 for <absolute path>
```

Reviewed By: dinhviethoa

fbshipit-source-id: 8be4ce1