Move artifact fetch service into the cache itself
Summary:
After recent refactorings, it's pretty straightforward to move this into
the cache itself.

This has has the side effect of moving dircache interaction back onto
the main build threads. The fact that the dircache stuff was on the
other service was mostly just because it was difficult to move the http
stuff there without also bringing along the dircache stuff.

Test Plan: CI, `buck build buck` and verify chrome trace looks correct

Reviewed By: yiding

fbshipit-source-id: 64d056c