Add functionality to fetch all source files during preload.
Summary:
This diff adds a config option for materializing source files on demand, which when set to
false, makes the slave materialize all source files with contents before the build actually starts.
This might not be a very good idea after we have distributed builds with each slave only building
a small portion of the build graph.

Test Plan: CI

Reviewed By: aiked

fbshipit-source-id: e7ca752