Fix one-level cache bug
Summary: Separate the conditionals to allow for storing of metadata and content at a single layer. Store metadata as long as it isn't empty, and store content if the metadata does not specifically contain the key indicating that this is the metadata-only level of a two-level cache.

Test Plan: `buck build` with cache.mode=sqlite,thrift_over_http

Reviewed By: asp2insp, bertmaher

fbshipit-source-id: 94558f1