Introduce ProjectFilesystemDelegate.getDetailsForLogging()
Summary:
This supersedes what was done in D5620740 to add more details in the
Chrome trace about the `ProjectFilesystemDelegate` that was used. In
particular, I worry about misspelling `eden.disable_sha1_fast_path`
and it not having the desired effect, so this is a simple way to verify
it.

Test Plan:
Do a Buck build in Eden with and without
`--config eden.disable_sha1_fast_path=true` and run the following after each
run and verify it has the expected output:

```
jq '[.[] | select(.name=="ProjectFilesystemDelegate")]' buck-out/log/build.trace
```

Reviewed By: ttsugriy

fbshipit-source-id: 6b54cca